Add shared BranchingNovelApp module
